refactor(api): split MeshConfiguration class into constant and service
