--- api/src/certification/results/domain/models/v3/CertificateLevels.js
@@ -1,11 +1,0 @@
-export const CERTIFICATE_LEVELS = {
-  preBeginner: 'LEVEL_PRE_BEGINNER',
-  beginner1: 'LEVEL_BEGINNER_1',
-  beginner2: 'LEVEL_BEGINNER_2',
-  independent3: 'LEVEL_INDEPENDENT_3',
-  independent4: 'LEVEL_INDEPENDENT_4',
-  advanced5: 'LEVEL_ADVANCED_5',
-  advanced6: 'LEVEL_ADVANCED_6',
-  expert7: 'LEVEL_EXPERT_7',
-  expert8: 'LEVEL_EXPERT_8',
-};
